Creating utilities, which contain functions to be used across the board that are not necessarily to be made public with the library
