+++ source/placentagen/generate_shapes.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 import numpy as np
-
+import pg_utilities
 
 def equispaced_data_in_ellipsoid(n, volume, thickness, ellipticity):
     # Generates equally sapced data points in an ellipsoid with the following inputs
@@ -78,7 +78,7 @@ def uniform_data_on_ellipsoid(n, volume, thickness, ellipticity, random_seed):
         if ((new_x / x_radius) ** 2 + (new_y / y_radius) ** 2) < 1:  # on the surface
             if (generated_seed) == 0:
                 generated_seed = generated_seed + 1
-                new_z = z_from_xy(new_x, new_y, x_radius, y_radius, z_radius)
+                new_z = pg_utilities.z_from_xy(new_x, new_y, x_radius, y_radius, z_radius)
                 chorion_data[generated_seed - 1][:] = [new_x, new_y, new_z]
             else:
                 reject = False
@@ -92,7 +92,7 @@ def uniform_data_on_ellipsoid(n, volume, thickness, ellipticity, random_seed):
                         break
                 if reject == False:
                     generated_seed = generated_seed + 1
-                    new_z = z_from_xy(new_x, new_y, x_radius, y_radius, z_radius)
+                    new_z = pg_utilities.z_from_xy(new_x, new_y, x_radius, y_radius, z_radius)
                     chorion_data[generated_seed - 1][:] = [new_x, new_y, new_z]
 
         attempts = attempts + 1
@@ -164,25 +164,4 @@ def calculate_ellipse_radii(volume, thickness, ellipticity):
     return {'x_radius': x_radius, 'y_radius': y_radius, 'z_radius': z_radius}
 
 
-def z_from_xy(x, y, x_radius, y_radius, z_radius):
-    z = z_radius * np.sqrt(1.0 - (x / x_radius) ** 2 - (y / y_radius) ** 2)
-    return z
 
-
-def check_in_ellipsoid(x, y, z, x_radius, y_radius, z_radius):
-    in_ellipsoid = False  # default to false
-    coord_check = (x / x_radius) ** 2 + (y / y_radius) ** 2 + (z / z_radius) ** 2
-    if coord_check < 1.0:
-        in_ellipsoid = True
-
-    return in_ellipsoid
-
-
-def check_on_ellipsoid(x, y, z, x_radius, y_radius, z_radius):
-    zero_tol = 1e-14
-    on_ellipsoid = False  # default to false
-    coord_check = (x / x_radius) ** 2 + (y / y_radius) ** 2 + (z / z_radius) ** 2
-    if abs(coord_check - 1.0) < zero_tol:
-        on_ellipsoid = True
-
-    return on_ellipsoid
