+++ source/placentagen/pg_utilities.py
@@ -1,0 +1,25 @@
+#!/usr/bin/env python
+import numpy as np
+
+def z_from_xy(x, y, x_radius, y_radius, z_radius):
+    z = z_radius * np.sqrt(1.0 - (x / x_radius) ** 2 - (y / y_radius) ** 2)
+    return z
+
+
+def check_in_ellipsoid(x, y, z, x_radius, y_radius, z_radius):
+    in_ellipsoid = False  # default to false
+    coord_check = (x / x_radius) ** 2 + (y / y_radius) ** 2 + (z / z_radius) ** 2
+    if coord_check < 1.0:
+        in_ellipsoid = True
+
+    return in_ellipsoid
+
+
+def check_on_ellipsoid(x, y, z, x_radius, y_radius, z_radius):
+    zero_tol = 1e-14
+    on_ellipsoid = False  # default to false
+    coord_check = (x / x_radius) ** 2 + (y / y_radius) ** 2 + (z / z_radius) ** 2
+    if abs(coord_check - 1.0) < zero_tol:
+        on_ellipsoid = True
+
+    return on_ellipsoid
